news_classifier/requirements.txt
tensorflow
numpy
scipy
nltk
sklearn
pandas
requests
watchdog
jsonrpclib-pelix
tap-news-utils
certifi
